# src/http/nodegen/interfaces/EventLocation.ts
export interface EventLocation {
  _id?: string;
  address?: string;
  city?: string;
  country?: string;
  latitude?: string;
  longitude?: string;
  name?: string;
  state?: string;
  zip?: string;
}